+++ src/Todo.md
@@ -15,11 +15,8 @@
 - add way to share app url
 - add way to share url with specific game settings
 
-## NGrid
+## Crossle
 
-- better name
-- make icon
-- change to default to home insted of ngrid
 - add way to move whole board or group of words
 
 ## Gribbles
