+++ src/crossle/components/Result.js
