+++ src/crossle/logic/gameSolvedQ.js
