+++ src/crossle/logic/getRegexForRow.js
